backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
azure-cosmos==4.7.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
